Roles + Login +  Basic Security
Adds Roles, Login and register and Basic security. Basic security currently is having a bug where when login button or register button is pressed it brings a login pop up. I am working on fixing it but it is an excruciating task that has little guides and takes a lot of time.